+++ src/matrixrtc/CallMembership.ts
@@ -16,7 +16,7 @@ limitations under the License.
 
 import { MXID_PATTERN } from "../models/room-member.ts";
 import { deepCompare } from "../utils.ts";
-import { isLivekitFocusSelection, type LivekitFocusSelection } from "./LivekitFocus.ts";
+import { isLivekitFocusSelection, type LivekitFocusSelection } from "./LivekitTransport.ts";
 import { slotDescriptionToId, slotIdToDescription, type SlotDescription } from "./MatrixRTCSession.ts";
 import type { RTCCallIntent, Transport } from "./types.ts";
 import { type MatrixEvent } from "../models/event.ts";
